TASK-2_3: Implement DDD in this project
