+++ lib/domain/blockchain/base_dto.dart
@@ -1,0 +1,4 @@
+class BaseResultDto {
+  String? errorMessage;
+  String timestamp = DateTime.now().millisecondsSinceEpoch.toString();
+}
