+++ lib/domain/blockchain/dto_collections.dart
