+++ lib/domain/blockchain/secret_dto.dart
