--- lib/infrastructure/blockchain/models/base_dto.dart
@@ -1,6 +1,0 @@
-import 'package:hnotes/util/common_data.dart';
-
-class BaseResultDto {
-  String? errorMessage;
-  String timestamp = DateTime.now().millisecondsSinceEpoch.toString();
-}
