+++ lib/infrastructure/blockchain/services/blockchain_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'package:http/http.dart' as http;
 
-import 'package:hnotes/util/common_data.dart';
-import 'package:hnotes/infrastructure/blockchain/models/dto_collections.dart';
+import 'package:hnotes/domain/common_data.dart';
+import 'package:hnotes/domain/blockchain/dto_collections.dart';
 import 'package:hnotes/infrastructure/blockchain/services/request_helper.dart';
 
 
@@ -12,17 +12,17 @@ final requestHeaders = {'Content-type': 'application/json'};
 class BlockchainRepository {
   final client = http.Client();
 
-  // Returns the number of the most recent block.
+  /// Returns the number of the most recent block.
   Future<Map<String, dynamic>> getLatestBlockNumber() async {
     return (await _parseNumberResultDto("eth_blockNumber")).toMap();
   }
 
-  // Returns the current price per gas in wei.
+  /// Returns the current price per gas in wei.
   Future<Map<String, dynamic>> getCurrentGasPrice() async {
     return (await _parseNumberResultDto("eth_gasPrice")).toMap();
   }
 
-  // Returns the current network name.
+  /// Returns the current network name.
   Future<Map<String, dynamic>> getNetwork() async {
     final NumberResultDto networkResult = await _parseNumberResultDto("net_version");
     final Map<String, String> networkNames = {
@@ -43,18 +43,18 @@ class BlockchainRepository {
     };
   }
 
-  // Returns the currently configured chain ID
+  /// Returns the currently configured chain ID
   Future<Map<String, dynamic>> getChainId() async {
     return (await _parseNumberResultDto("eth_chainId")).toMap();
   }
 
-  // Returns the current client version.
+  /// Returns the current client version.
   Future<Map<String, dynamic>> getNodeClientVersion() async {
     return (await _parseTextResult("web3_clientVersion")).toMap();
   }
 
 
-  // For API responses that only contain a number (hex or decimal)
+  /// For API responses that only contain a number (hex or decimal)
   Future<NumberResultDto> _parseNumberResultDto(String methodName) async {
     final String requestBody = formRequestBody(methodName);
 
@@ -65,7 +65,7 @@ class BlockchainRepository {
     });
   }
 
-  // For API responses that only contain a text string
+  /// For API responses that only contain a text string
   Future<TextResultDto> _parseTextResult(String methodName) async {
     final String requestBody = formRequestBody(methodName);
 
@@ -75,6 +75,4 @@ class BlockchainRepository {
       return dto;
     });
   }
-
 }
-
