+++ lib/infrastructure/blockchain/services/request_helper.dart
@@ -4,11 +4,11 @@ import 'package:yaml/yaml.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/services.dart' show rootBundle;
 
-import 'package:hnotes/util/common_data.dart';
-import 'package:hnotes/infrastructure/blockchain/models/secret_dto.dart';
+import 'package:hnotes/domain/blockchain/secret_dto.dart';
 
 List defaultParameter = [];
 final requestHeaders = {'Content-type': 'application/json'};
+final String secretsFilePath = "assets/secrets/secret.yaml";
 
 String formRequestBody(String method, [var parameter]) {
   parameter ??= defaultParameter;
