+++ lib/infrastructure/local_storage/start_day/start_day_repository.dart
@@ -1,0 +1,9 @@
+import 'package:hnotes/infrastructure/local_storage/share_preferences.dart';
+
+class StartDayRepository {
+  /// Get saved love start date API
+  Future<String> getLoveStartDate() async {
+    String theDate = await getDataFromSharedPref('startDate');
+    return theDate;
+  }
+}
