--- lib/requester/repository.dart
@@ -1,11 +1,0 @@
-import 'package:hnotes/requester/api_provider.dart';
-import 'package:hnotes/note_services/note_model.dart';
-
-class Repository {
-  final apiProvider = NoteApiProvider();
-
-  Future<NoteModel> fetchAllNotes() => apiProvider.getAllNotes();
-
-  Future<String> fetchLoveStartDate() => apiProvider.getLoveStartDate();
-
-}
