--- lib/splash_screen/splash_collections.dart
@@ -1,2 +1,0 @@
-export 'count_day_bloc.dart';
-export 'count_day_model.dart';
