--- lib/util/util_collections.dart
@@ -1,2 +1,0 @@
-export '../components/divider.dart';
-export '../components/size_box.dart';
